@@ -2,7 +2,7 @@
   "name": "danabrey/espn-college-football-stats",
   "description": "PHP package for reading data from ESPN's college football JSON API",
   "type": "library",
-  "version": "1.1.1",
+  "version": "2.0",
   "license": "MIT",
   "authors": [
     {
@@ -11,7 +11,7 @@
     }
   ],
   "require": {
-    "php": "^7.1",
+    "php": "^7.4",
     "symfony/serializer": "^4.3",
     "symfony/property-access": "^4.3",
     "ext-json": "*"
