@@ -3,45 +3,10 @@ namespace DanAbrey\ESPNCollegeFootballStats;
 
 class Team
 {
-    /**
-     * @var string
-     */
-    private $id;
+    public string $id;
 
     /**
-     * @var array
+     * @var TeamSeason[]|array
      */
-    private $seasons = [];
-
-    /**
-     * @return string
-     */
-    public function getId(): string
-    {
-        return $this->id;
-    }
-
-    /**
-     * @param string $id
-     */
-    public function setId(string $id): void
-    {
-        $this->id = $id;
-    }
-
-    /**
-     * @return array
-     */
-    public function getSeasons(): array
-    {
-        return $this->seasons;
-    }
-
-    /**
-     * @param array $seasons
-     */
-    public function setSeasons(array $seasons): void
-    {
-        $this->seasons = $seasons;
-    }
+    public array $seasons = [];
 }
